updated items class doubles to BigDecimals along with child classes, added comments for sales report class for flow. removed comment from top of item class that was nto needed

+++ src/main/java/com/techelevator/snacks/Candy.java
@@ -1,7 +1,9 @@
 package com.techelevator.snacks;
 
+import java.math.BigDecimal;
+
 public class Candy extends Items{
-    public Candy( String position, String itemName, int quantity, double cost ){
+    public Candy(String position, String itemName, int quantity, BigDecimal cost ){
         super(position,itemName,quantity,cost);
         this.setFlavourText("Munch Munch, Yum!");
     }
